backend/requirements.txt
alembic==1.4.3
aniso8601==8.0.0
attrs==20.2.0
click==7.1.2
coverage==5.3
Flask==1.1.2
flask-marshmallow==0.14.0
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
importlib-metadata==2.0.0
iniconfig==1.0.1
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.8.0
marshmallow-sqlalchemy==0.23.1
packaging==20.4
pg8000==1.16.6
pluggy==0.13.1
psycopg2==2.8.6
py==1.9.0
pyparsing==2.4.7
pytest==6.1.1
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
scramp==1.2.0
six==1.15.0
SQLAlchemy==1.3.19
testing.common.database==2.0.3
testing.postgresql==1.3.0
toml==0.10.1
Werkzeug==1.0.1
zipp==3.3.0
